Move coordinates.hpp implementation to coordinates.cpp, menus are now in menu.cpp
